% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmtp-package.R
\docType{package}
\name{hmtp-package}
\alias{hmtp}
\alias{hmtp-package}
\title{hmtp: Targeted Minimum-Loss Based Estimation for Semi-continuous Outcomes}
\description{

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mtpverse/hmtp}
  \item Report bugs at \url{https://github.com/mtpverse/hmtp/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Williams \email{ntwilliams.personal@gmail.com} (\href{https://orcid.org/0000-0002-1378-4831}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Iván Díaz \email{ild2005@med.cornell.edu} (\href{https://orcid.org/0000-0001-9056-2047}{ORCID}) [copyright holder]
}

}
\keyword{internal}
