% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{hmtp_contrast}
\alias{hmtp_contrast}
\title{Perform Contrasts of HMTP Fits}
\usage{
hmtp_contrast(..., ref)
}
\arguments{
\item{...}{One or more objects of class 'hmtp'.}

\item{ref}{A reference value or another 'hmtp' fit to compare all other fits against.}
}
\value{
A list of class \code{hmtp_contrast} containing the following components:

\item{type}{The type of contrast performed.}
\item{null}{The null hypothesis.}
\item{vals}{A dataframe containing the contrasts estimates, standard errors, and confidence intervals.}
\item{eifs}{Un-centered estimated influence functions for contrasts estimated.}
}
\description{
Estimates contrasts of multiple HMTP fits compared to either a known reference value
or a reference HMTP fit.
}
