% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{hmtp_tmle}
\alias{hmtp_tmle}
\title{HMTP Targeted Maximum Likelihood Estimator}
\usage{
hmtp_tmle(
  data,
  trt,
  outcome,
  baseline = NULL,
  cens = NULL,
  shift = NULL,
  shifted = NULL,
  mtp = FALSE,
  id = NULL,
  upper_bound = NULL,
  learners_trt = "SL.glm",
  learners_zero = "SL.glm",
  learners_positive = "SL.glm",
  folds = 10,
  weights = NULL,
  log = TRUE,
  control = hmtp_control(),
  ...
)
}
\arguments{
\item{data}{\[\code{data.frame}\]\cr
A \code{data.frame} in wide format containing all necessary variables
for the estimation problem. Must not be a \code{data.table}.}

\item{trt}{\[\code{character}\]\cr
The column name of the treatment variable.}

\item{outcome}{\[\code{character}\]\cr
The column name of the outcome variable.}

\item{baseline}{\[\code{character}\]\cr
An optional vector containing the column names of baseline covariates to be
included for adjustment.}

\item{cens}{\[\code{character}\]\cr
An optional vector for the column name of a censoring indicator. If missingness in the outcome is
present this must be provided.}

\item{shift}{\[\code{closure}\]\cr
A two argument function that specifies how the treatment variable should be shifted.
See examples for how to specify shift functions for continuous, binary, and categorical exposures.}

\item{shifted}{\[\code{data.frame}\]\cr
An optional data frame, the same as in \code{data}, but modified according
to the treatment policy of interest. If specified, \code{shift} is ignored.}

\item{mtp}{\[\code{logical(1)}\]\cr
Is the intervention of interest a modified treatment policy?
Default is \code{FALSE}. If treatment variables are continuous this should be \code{TRUE}.}

\item{id}{\[\code{character(1)}\]\cr
An optional column name containing cluster level identifiers.}

\item{upper_bound}{\[\code{numeric(1)}\]\cr}

\item{learners_trt}{\[\code{character}\]\cr A vector of \code{SuperLearner} algorithms for estimation
of the exposure mechanism. Default is \code{"SL.glm"}, a main effects GLM.
\bold{Only include candidate learners capable of binary classification}.}

\item{learners_zero}{\[\code{character}\]\cr}

\item{learners_positive}{\[\code{character}\]\cr}

\item{folds}{\[\code{integer(1)}\]\cr
The number of folds to be used for cross-fitting.}

\item{weights}{\[\code{numeric(nrow(data))}\]\cr
An optional vector containing sampling weights.}

\item{...}{Extra arguments. Exists for backwards compatibility.}
}
\value{
A list of class \code{hmtp} containing the following components:
}
\description{
HMTP Targeted Maximum Likelihood Estimator
}
\details{
## Should \code{mtp = TRUE}?
A modified treatment policy (MTP) is an intervention that depends
on the natural value of the exposure (the value that the treatment would have taken under no intervention).
This differs from other causal effects,
such as the average treatment effect (ATE), where an exposure would be increased (or decreased) deterministically.
\bold{If your intervention of interest adds, subtracts, or multiplies the observed treatment values
by some amount, use \code{mtp = TRUE}}.
}
