% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmtp-package.R
\docType{package}
\name{htmp-package}
\alias{htmp}
\alias{htmp-package}
\title{htmp: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
