% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{static_binary_off}
\alias{static_binary_off}
\title{Turn All Treatment Nodes Off}
\usage{
static_binary_off(data, trt)
}
\arguments{
\item{data}{A \code{data.frame} containing the treatment variables.}

\item{trt}{The name of the current treatment variable.}
}
\value{
A A \code{data.frame} with all treatment nodes set to 0.
}
\description{
A pre-packaged shift function for use with provided estimators when the exposure is binary.
Used to estimate the population intervention effect when the treatment variable is set to 0.
}
\seealso{
[hmtp_tmle()]
}
