% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{static_binary_on}
\alias{static_binary_on}
\title{Turn All Treatment Nodes On}
\usage{
static_binary_on(data, trt)
}
\arguments{
\item{data}{A A \code{data.frame} containing the treatment variables.}

\item{trt}{The name of the current treatment variable.}
}
\value{
A \code{data.frame} with all treatment nodes set to 1.
}
\description{
A pre-packaged shift function for use with provided estimators when the exposure is binary.
Used to estimate the population intervention effect when the treatment variable i set to 1.
}
\seealso{
[hmtp_tmle()]
}
