% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.lmtp}
\alias{tidy.lmtp}
\title{Tidy a(n) lmtp object}
\usage{
\method{tidy}{lmtp}(x, ...)
}
\arguments{
\item{x}{A `lmtp` object produced by a call to [lmtp::lmtp_tmle()], [lmtp::lmtp_sdr()],
[lmtp::lmtp_sub()], or [lmtp::lmtp_ipw()].}

\item{...}{Unused, included for generic consistency only.}
}
\description{
Tidy a(n) lmtp object
}
\examples{
\donttest{
a <- c("A1", "A2")
nodes <- list(c("L1"), c("L2"))
cens <- c("C1", "C2")
y <- "Y"
fit <- lmtp_tmle(sim_cens, a, y, time_vary = nodes, cens = cens, shift = NULL, folds = 2)
tidy(fit)
}

}
