% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.hmtp}
\alias{tidy.hmtp}
\title{Tidy a(n) hmtp object}
\usage{
\method{tidy}{hmtp}(x, ...)
}
\arguments{
\item{x}{A \code{hmtp} object produced by a call to \code{\link[=hmtp_tmle]{hmtp_tmle()}}.}

\item{...}{Unused, included for generic consistency only.}
}
\description{
Tidy a(n) hmtp object
}
\examples{
\donttest{
}

}
